changed primary key to long

+++ backendParent/api/src/main/java/com/backend/api/admin/controller/CouponController.java
@@ -71,7 +71,7 @@ public class CouponController {
 													  @RequestParam(value = "couponId", required = true) String couponId) {
 		GenericResponse<String> response = new GenericResponse<String>();
 		try {
-			couponService.toggleCoupon(CommonUtil.isValidStringParam(couponId)? Integer.parseInt(couponId) : 0);
+			couponService.toggleCoupon(CommonUtil.isValidStringParam(couponId)? Long.parseLong(couponId) : 0);
 			response.setStatus(Response.Status.OK);
 		} catch (Exception ex) {
 			logger.error("toggleCouponStatus : " + ex);
@@ -84,7 +84,7 @@ public class CouponController {
 	
 	@RequestMapping(value = "/deleteCoupon", method = RequestMethod.DELETE, produces = MediaType.APPLICATION_JSON_VALUE)
 	public GenericResponse<String> deleteCoupon(HttpServletRequest request,
-											    @RequestParam(value = "couponId", required = true) int couponId) {
+											    @RequestParam(value = "couponId", required = true) Long couponId) {
 		GenericResponse<String> response = new GenericResponse<String>();
 		try {
 			couponService.deleteCoupon(couponId);
