+++ backendParent/persistence/src/main/java/com/backend/persistence/entity/CustomerAddress.java
@@ -32,7 +32,7 @@ public class CustomerAddress implements Serializable {
 	@Id
 	@GeneratedValue(strategy = GenerationType.AUTO)
 	@Column(name = "ADDRESSID")
-	private int addressId;
+	private Long addressId;
 
 	@Column(name = "DOORNUMBER")
 	private String doorNumber;
@@ -65,11 +65,11 @@ public class CustomerAddress implements Serializable {
 		this.tenant = tenant;
 	}
 
-	public int getAddressId() {
+	public Long getAddressId() {
 		return addressId;
 	}
 
-	public void setAddressId(int addressId) {
+	public void setAddressId(Long addressId) {
 		this.addressId = addressId;
 	}
