+++ backendParent/persistence/src/main/java/com/backend/persistence/repository/OrdersRepository.java
@@ -15,7 +15,7 @@ import com.backend.persistence.entity.Orders;
  *
  */
 @Repository
-public interface OrdersRepository extends JpaRepository<Orders, Integer> {
+public interface OrdersRepository extends JpaRepository<Orders, Long> {
 	
 	String findOrdersByIdQuery = "select order from Orders order where order.tenant = :tenant and order.orderId = :orderId";
 	String findOrdersForTenantQuery = "select order from Orders order where order.tenant = :tenant";
@@ -25,10 +25,10 @@ public interface OrdersRepository extends JpaRepository<Orders, Integer> {
 	String findCustomerOrdersQuery = "select * from orders where tenantid = ?1 and customerid= ?2";
 	
 	@Query(findOrdersByIdQuery)
-	Orders findOrdersById(@Param("tenant") Tenant realm, @Param("orderId") int orderId);
+	Orders findOrdersById(@Param("tenant") Tenant realm, @Param("orderId") Long orderId);
 	
 	@Query(findOrdersByCustomerQuery)
-	List<Orders> findOrdersByCustomer(@Param("tenant") Tenant realm, @Param("customerId") int customerId);
+	List<Orders> findOrdersByCustomer(@Param("tenant") Tenant realm, @Param("customerId") Long customerId);
 	
 	@Query(findOrdersForTenantQuery)
 	List<Orders> findOrdersForTenant(@Param("tenant") Tenant realm);
@@ -37,9 +37,9 @@ public interface OrdersRepository extends JpaRepository<Orders, Integer> {
 	List<Orders> findLimitedOrders(String tenant, int limit, int offset);
 	
 	@Query(value = findCustomerOrdersQuery, nativeQuery = true)
-	List<Orders> findCustomerOrders(String tenant, int customerId);
+	List<Orders> findCustomerOrders(String tenant, Long customerId);
 	
 	@Query(findOrdersByStatusForEmployeeQuery)
-	List<Orders> findOrdersByStatusForEmployee(@Param("tenant") Tenant realm, @Param("employeeId") int employeeId,  @Param("status") String string);
+	List<Orders> findOrdersByStatusForEmployee(@Param("tenant") Tenant realm, @Param("employeeId") Long employeeId,  @Param("status") String string);
 
 }
