+++ backendParent/persistence/src/main/java/com/backend/persistence/repository/ProductImagesRepository.java
@@ -19,7 +19,7 @@ import com.backend.persistence.entity.ProductImages;
  *
  */
 @Repository
-public interface ProductImagesRepository extends JpaRepository<ProductImages, Integer> {
+public interface ProductImagesRepository extends JpaRepository<ProductImages, Long> {
 	
 	String findAllImagesForProductQuery = "select pi from ProductImages pi where pi.tenant = :tenant and pi.productId = :pId";
 	String findAllImagesForProductByIdQuery = "select * from ProductImages where tenantid = :tenant and productid = :pId";
@@ -31,15 +31,15 @@ public interface ProductImagesRepository extends JpaRepository<ProductImages, In
 	List<ProductImages> findAllImagesForProduct(@Param("tenant") Tenant realm, @Param("pId") Product product);
 	
 	@Query(value = findAllImagesForProductByIdQuery, nativeQuery = true)
-	List<ProductImages> findAllImagesForProduct(@Param("tenant") String realm, @Param("pId") int pId);
+	List<ProductImages> findAllImagesForProduct(@Param("tenant") String realm, @Param("pId") Long pId);
 	
 	@Query(findImageByIdQuery)
-	ProductImages findImageById(@Param("tenant") Tenant realm, @Param("imageId") int imageId);
+	ProductImages findImageById(@Param("tenant") Tenant realm, @Param("imageId") Long imageId);
 	
 	@Modifying
 	@Cascade(CascadeType.DELETE)
 	@Query(deleteByIdQuery)
-	void deleteById(@Param("tenant") Tenant realm, @Param("imageId") int imageId);
+	void deleteById(@Param("tenant") Tenant realm, @Param("imageId") Long imageId);
 	
 	@Query(getProductCountByIdQuery)
 	int getProductCount(@Param("tenant") Tenant realm, @Param("pId") Product product);
