+++ backendParent/persistence/src/main/java/com/backend/persistence/service/CouponsService.java
@@ -16,11 +16,11 @@ public interface CouponsService {
 
 	List<Coupons> findAllCouponsForTenant();
 
-	void toggleCoupon(int id);
+	void toggleCoupon(Long id);
 
-	void deleteCoupon(int id);
+	void deleteCoupon(Long id);
 
-	Coupons findCouponById(int id);
+	Coupons findCouponById(Long id);
 
 	Coupons getCouponByCode(String code);
