+++ backendParent/persistence/src/main/java/com/backend/persistence/service/OrdersService.java
@@ -19,7 +19,7 @@ public interface OrdersService {
 
 	List<Orders> getOrders(int limit, int offset);
 
-	void updateOrderStatus(String status, int orderId) throws Exception;
+	void updateOrderStatus(String status, Long orderId) throws Exception;
 
 	int getAllUnassignedOrdersCount() throws Exception;
 
@@ -31,6 +31,6 @@ public interface OrdersService {
 
 	List<Orders> getCustomerOrders();
 
-	void createCustomerOrder(int couponId, int paymentMode, int addressId, int deliveryCharge) throws Exception;
+	void createCustomerOrder(Long couponId, int paymentMode, Long addressId, int deliveryCharge) throws Exception;
 
 }
